+++ src/main/java/org/jabref/gui/entryeditor/DeprecatedFieldsTab.java
@@ -25,6 +25,8 @@ import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.entry.field.Field;
 import org.jabref.preferences.PreferencesService;
 
+import com.tobiasdiez.easybind.EasyBind;
+
 public class DeprecatedFieldsTab extends FieldsEditorTab {
 
     private final BibEntryTypesManager entryTypesManager;
@@ -44,7 +46,13 @@ public class DeprecatedFieldsTab extends FieldsEditorTab {
         this.entryTypesManager = entryTypesManager;
 
         setText(Localization.lang("Deprecated fields"));
-        setTooltip(new Tooltip(Localization.lang("Show deprecated BibTeX fields")));
+        EasyBind.subscribe(preferences.getGeneralPreferences().showAdvancedHintsProperty(), advancedHints -> {
+            if (advancedHints) {
+                setTooltip(new Tooltip(Localization.lang("Shows fields having a successor in biblatex.\nFor instance, the publication month should be part of the date field.\nUse the Cleanup Entries functionality to convert the entry to biblatex.")));
+            } else {
+                setTooltip(new Tooltip(Localization.lang("Shows fields having a successor in biblatex.")));
+            }
+        });
         setGraphic(IconTheme.JabRefIcons.OPTIONAL.getGraphicNode());
     }
 
@@ -52,7 +60,7 @@ public class DeprecatedFieldsTab extends FieldsEditorTab {
     protected Set<Field> determineFieldsToShow(BibEntry entry) {
         Optional<BibEntryType> entryType = entryTypesManager.enrich(entry.getType(), databaseContext.getMode());
         if (entryType.isPresent()) {
-            return entryType.get().getDeprecatedFields().stream().filter(field -> !entry.getField(field).isEmpty()).collect(Collectors.toSet());
+            return entryType.get().getDeprecatedFields(databaseContext.getMode()).stream().filter(field -> !entry.getField(field).isEmpty()).collect(Collectors.toSet());
         } else {
             // Entry type unknown -> treat all fields as required
             return Collections.emptySet();